+++ ProyectoG1/Models/TipoInstitucion.cs
@@ -20,7 +20,7 @@ namespace ProyectoG1.Models
             this.Institucion = new HashSet<Institucion>();
         }
     
-        public int IdTipoInstitucion { get; set; }
+        public long IdTipoInstitucion { get; set; }
         public string Nombre { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
